# backend/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Test Tracker API",
    "version": "1.0.0",
    "description": "Documentazione completa delle API del progetto Test Tracker per supportare il frontend e il test delle route. Le API protette richiedono il token JWT nell'header Authorization come Bearer token."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Server locale di sviluppo"
    }
  ],
...
  "paths": {
...
    "/api/auth/bootstrap": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Crea il primissimo account superadmin (pubblica, si autodisabilita)",
        "description": "Rotta pubblica (nessun token richiesto) pensata per un deploy o un database appena creato, dove nessun utente esiste ancora e /register non è raggiungibile perché richiede già un token superadmin. Crea un utente con ruolo superadmin fisso e la password scelta direttamente (a differenza di /register, qui non c'è alcun invio email/token di setup). Funziona solo se la tabella utenti è vuota: risponde 403 non appena esiste almeno un utente, anche creato manualmente via SQL diretto.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "surname": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 }
                },
                "required": ["name", "surname", "email", "password"]
              },
              "example": {
                "name": "Mario",
                "surname": "Rossi",
                "email": "mario.rossi@example.com",
                "password": "unapasswordsicura"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account superadmin creato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "nome": { "type": "string" },
                        "cognome": { "type": "string" },
                        "email": { "type": "string" },
                        "role": { "type": "string", "enum": ["superadmin"] }
                      }
                    }
                  }
                },
                "example": {
                  "message": "Account superadmin creato con successo.",
                  "user": {
                    "id": 1,
                    "nome": "Mario",
                    "cognome": "Rossi",
                    "email": "mario.rossi@example.com",
                    "role": "superadmin"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Campi mancanti, password troppo corta o email già esistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "examples": {
                  "missingFields": {
                    "value": {
                      "message": "Nome, cognome, email e password sono obbligatori."
                    }
                  },
                  "passwordTooShort": {
                    "value": {
                      "message": "La password deve contenere almeno 8 caratteri."
                    }
                  },
                  "emailExists": {
                    "value": {
                      "error": "Richiesta non valida",
                      "message": "L'utenza con la mail mario.rossi@example.com esiste già."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Bootstrap non disponibile: esiste già almeno un utente nel sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Bootstrap non disponibile: esiste già almeno un utente nel sistema."
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetail"
                },
                "example": {
                  "message": "Errore del server",
                  "specific": "Internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registra un nuovo utente (solo superadmin)",
        "description": "Crea un nuovo utente con nome, cognome, email e ruolo, senza richiedere una password. Genera un token di setup monouso (48h) e invia una email con il link per impostare la password (POST /api/auth/setup). Riservata al superadmin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "name": "Mario",
                "surname": "Rossi",
                "email": "mario.rossi@example.com",
                "role": "user"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utente creato, email di configurazione inviata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                },
                "example": {
                  "message": "Utente creato. Email di configurazione inviata."
                }
              }
            }
          },
          "400": {
            "description": "Campi mancanti, ruolo non valido o email già esistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "examples": {
                  "missingFields": {
                    "value": {
                      "message": "Nome, cognome, email e ruolo sono obbligatori."
                    }
                  },
                  "invalidRole": {
                    "value": {
                      "message": "Ruolo non valido"
                    }
                  },
                  "emailExists": {
                    "value": {
                      "error": "Richiesta non valida",
                      "message": "L'utenza con la mail mario.rossi@example.com esiste già."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Permessi insufficienti (solo superadmin)"
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetail"
                },
                "example": {
                  "message": "Errore del server",
                  "specific": "Duplicate entry 'mario.rossi@example.com' for key 'email'"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Richiede il reset della password",
        "description": "Rotta pubblica: la risposta è identica indipendentemente dal fatto che l'email esista o meno, per non rivelare quali indirizzi sono registrati. Se l'email corrisponde a un utente, invalida gli eventuali token di setup/reset non ancora usati, ne genera uno nuovo e invia un'email con il link per impostare la password (stessa rotta di POST /api/auth/setup). L'email è di tipo 'setupPassword' se l'utente non ha mai impostato una password, 'resetPassword' altrimenti.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "mario.rossi@example.com"
                  }
                },
                "required": [
                  "email"
                ]
              },
              "example": {
                "email": "mario.rossi@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Richiesta accettata (indipendentemente dall'esistenza dell'email)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                },
                "example": {
                  "message": "Se l'email è registrata, riceverai a breve un'email con le istruzioni."
                }
              }
            }
          },
          "400": {
            "description": "Email mancante",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "L'email è obbligatoria."
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetail"
                },
                "example": {
                  "message": "Errore del server",
                  "specific": "Internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/verify-setup-token": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Verifica la validità di un token di setup/reset senza consumarlo",
        "description": "Rotta pubblica usata dalla pagina /setup per capire se mostrare il form o un messaggio di link non valido (es. quando l'utente torna indietro dopo aver già impostato la password). Non marca il token come usato.",
        "security": [],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "3f1a9c7e2b8d4f6a1c9e0b2d7f4a6c8e1b3d5f7a9c1e3b5d7f9a1c3e5b7d9f1a"
          }
        ],
        "responses": {
          "200": {
            "description": "Token valido (non usato, non scaduto)",
            "content": {
              "application/json": {
                "example": {
                  "valid": true
                }
              }
            }
          },
          "404": {
            "description": "Token mancante, non valido, scaduto o già usato",
            "content": {
              "application/json": {
                "examples": {
                  "missingToken": {
                    "value": {
                      "valid": false,
                      "message": "Link non valido."
                    }
                  },
                  "invalidToken": {
                    "value": {
                      "valid": false,
                      "message": "Link non valido o già utilizzato."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetail"
                },
                "example": {
                  "message": "Errore del server",
                  "specific": "Internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/setup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Completa la configurazione dell'account tramite token",
        "description": "Verifica il token di setup ricevuto via email (hash SHA-256, non ancora usato, non scaduto), imposta la password dell'utente e marca il token come usato. Rotta pubblica, non richiede autenticazione.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupAccountRequest"
              },
              "example": {
                "token": "3f1a9c7e2b8d4f6a1c9e0b2d7f4a6c8e1b3d5f7a9c1e3b5d7f9a1c3e5b7d9f1a",
                "password": "Password123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password impostata con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                },
                "example": {
                  "message": "Password impostata con successo. Puoi ora effettuare il login."
                }
              }
            }
          },
          "400": {
            "description": "Campi mancanti, token non valido, già usato o scaduto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "examples": {
                  "missingFields": {
                    "value": {
                      "message": "Token e password sono obbligatori."
                    }
                  },
                  "invalidToken": {
                    "value": {
                      "message": "Link non valido o scaduto."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetail"
                },
                "example": {
                  "message": "Errore del server",
                  "specific": "Internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica un utente",
        "description": "Verifica email e password e restituisce un token JWT insieme ai dati utente. Questa route è pubblica e non richiede token in ingresso.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "mario.rossi@example.com",
                "password": "Password123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login effettuato correttamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                },
                "example": {
                  "message": "Login utente",
                  "user": {
                    "id": 1,
                    "name": "Mario",
                    "surname": "Rossi",
                    "email": "mario.rossi@example.com",
                    "role": "user"
                  },
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "400": {
            "description": "Campi mancanti",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Email e password sono obbligatori"
                }
              }
            }
          },
          "401": {
            "description": "Password non valida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Password non valida"
                }
              }
            }
          },
          "403": {
            "description": "Account non ancora configurato (password non impostata tramite il link di setup)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Account non ancora configurato. Controlla la tua email."
                }
              }
            }
          },
          "404": {
            "description": "Utente non trovato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Utente non trovato"
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetail"
                },
                "example": {
                  "message": "Errore del server",
                  "specific": "Internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Effettua il logout",
        "description": "Logout documentato come operazione protetta da token JWT.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout effettuato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                },
                "example": {
                  "message": "Logout effettuato con successo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Inserire il token JWT come: Bearer <token>"
      }
    },
    "schemas": {
...
      "SuccessMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ErrorWithDetail": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "specific": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mario"
          },
          "surname": {
            "type": "string",
            "example": "Rossi"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "mario.rossi@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin",
              "superadmin"
            ],
            "example": "user"
          }
        },
        "required": [
          "name",
          "surname",
          "email",
          "role"
        ]
      },
      "SetupAccountRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token di setup ricevuto via email (parametro ?token= nel link)",
            "example": "3f1a9c7e2b8d4f6a1c9e0b2d7f4a6c8e1b3d5f7a9c1e3b5d7f9a1c3e5b7d9f1a"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Minimo 8 caratteri",
            "example": "Password123!"
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "mario.rossi@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Password123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Mario"
          },
          "cognome": {
            "type": "string",
            "example": "Rossi"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "mario.rossi@example.com"
          },
          "role": {
            "type": "string",
            "example": "user"
          }
        },
        "required": [
          "id",
          "nome",
          "cognome",
          "email",
          "role"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login utente"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "token": {
            "type": "string",
            "description": "JWT da inviare nell'header Authorization delle route protette",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "message",
          "user",
          "token"
        ]
      },
...
    }
  }
}